requirement.txt
builtwith==1.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pyfiglet==0.8.post1
python-nmap==0.7.1
requests==2.30.0
six==1.16.0
termcolor==2.3.0
urllib3==2.0.2
